Hammer: refactor for clearer separation
This is just moving code around for clearer separation and documentation
via method naming.
